backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
torch
torchvision>=0.24.0
Pillow>=11.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
geopy==2.4.1
scikit-learn>=1.6.0
passlib[bcrypt]==1.7.4
easyocr==1.7.1
